Xong giỏ hàng
- thêm, xóa, sửa sản phẩm trong giỏ hàng
- tạo đơn hàng
- lưu đơn hàng/chi tiết đơn hàng vào csdl

+++ src/shop/entity/OderDetail.java
@@ -22,7 +22,7 @@ public class OderDetail implements Serializable{
 	private Product product;
 	
 	private Integer Quantity;
-	private Float UnitPrice;
+	private Double UnitPrice;
 	
 	public Oder getOder() {
 		return oder;
@@ -42,10 +42,10 @@ public class OderDetail implements Serializable{
 	public void setQuantity(Integer quantity) {
 		Quantity = quantity;
 	}
-	public Float getUnitPrice() {
+	public Double getUnitPrice() {
 		return UnitPrice;
 	}
-	public void setUnitPrice(Float unitPrice) {
+	public void setUnitPrice(Double unitPrice) {
 		UnitPrice = unitPrice;
 	}
